dom/base/test/chrome/test_bug780199.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=780199
-->
<window title="Mozilla Bug 780199"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="test()">
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=780199"
     target="_blank">Mozilla Bug 780199</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript">
  <![CDATA[

  /** Test for Bug 780199 **/

  SimpleTest.waitForExplicitFinish();

  var b;

  function callback(r) {
    is(r[0].type, "attributes");
    is(r[0].oldValue, b.getAttribute("src"));
    setTimeout(continueTest, 500);
  }

  function continueTest() {
    // Check that a new page wasn't loaded.
    is(b.contentDocument.documentElement.textContent, "testvalue");
    SimpleTest.finish();
  }

  function test() {
    b = document.getElementById("b");
    var m = new MutationObserver(callback);
    m.observe(b, { attributes: true, attributeOldValue: true });
    b.contentDocument.documentElement.textContent = "testvalue";
    b.setAttribute("src", b.getAttribute("src"));
  }

  ]]>
  </script>
  <browser id="b" src="data:text/plain,initial"/>
</window>
